Layer 3 Step 12.3d: Preserve fractional pCore counts in licensing math — drift 55 -> 37
ROOT CAUSE
==========
The engine's WorkloadInventory model declared pCore counts as ``int``,
which caused fractional values from the BA workbook to be truncated:

  client.pcores_with_sql_server_esu = 78.85135 (formula =1167/1.48)
                                   -> 78 (engine, after int cast)
                                   -> $4,766 SQL ESU drift per year (~1%)

The BA workbook 1-Client Variables sheet (Customer A) uses fractional
vCPU/vcpu_per_pcore divisions throughout:

  D67  pcores_with_windows_server  = =12405/D66 = 8381.756...
  D68  pcores_with_windows_esu     =  =1167/D66 =  788.513...
  D70  pcores_with_sql_server      =  =D67*0.1  =  838.175...
  D71  pcores_with_sql_server_esu  =  =D68*0.1  =   78.851...

The replica oracle (layer3_status_quo.compute_baselines) applies the
per-core licensing rate to the fractional value verbatim. Engine
silently truncated to int, producing systematic ~1% understatement on
all license cells.

FIX
===
Change pcores_with_* fields from int to float end-to-end:

1. engine/models.py WorkloadInventory:
   - 5 fields: pcores_with_virtualization, pcores_with_windows_server,
     pcores_with_windows_esu, pcores_with_sql_server,
     pcores_with_sql_esu -> float
   - derive_defaults validator: removed 3 round() calls

2. engine/rvtools_parser.py:
   - 4 struct fields on RVToolsInventory: int -> float
   - 6 round() calls in pCore derivation removed (preserves fractional
     vcpus / vcpu_per_core_ratio result)

3. training/replicas/engine_bridge_l3.py:
   - 5 int(client.pcores_with_*) casts -> float(...)

ADVERSARIAL REVIEW
==================
Explore subagent (thorough mode) found 5 breaking changes vs. the
3-change minimum proposal:

- Required: also update RVTools struct + remove 6 round() calls there
- Required: also update bridge int() -> float()
- test_engine.py line 41 (== 543) safe: contoso fixture passes
  explicit 543, 543 == 543.0 is True in Python
- No int-arithmetic licensing tier breakpoints downstream
- All consumers (status_quo, financial_case, retained_costs, etc.)
  accept float naturally
- Division guards (max(..., 1)) already in place

VERIFIED
========
- Replica still 395/395 CLEAN.
- Engine drift: 55 -> 37 (18 cells fixed in one shot).
- All 11 status_quo.SQL Server ESU.Y* cells PASS.
- cash_flow.Savings (SQ-AZ) cells: 5 -> 2 (cascade healed).
- cash_flow.CF Delta (AZ-SQ) cells: 5 -> 2.
- 60/70 tests pass (10 pre-existing RVtools/consumption failures
  unrelated, unchanged).

REMAINING DRIFT (37) DOMINATED BY
=================================
- cash_flow.CF Rate.Y0..Y10 (10 cells; ratio with strict tolerance
  cascading from sub-$5K absolute deltas)
- status_quo.NW+Fitout Depreciation.Y5..Y10 (6 sub-$1.10 cells;
  fractional-host residual; deferred)
- cash_flow.AZ Migration.Y1, Y2 (2 cells)
- cash_flow.Savings (SQ-AZ).Y0, Y2 (2)
- cash_flow.CF Delta (AZ-SQ).Y0, Y2 (2)
- Headlines / payback (~13 cells, NPV cascades)

+++ training/replicas/engine_bridge_l3.py
@@ -160,11 +160,14 @@ def replica_inputs_to_engine_inputs(
         vm_to_server_ratio=bm.vm_to_physical_server_ratio,
         byol_virtualization_for_avs=_yesno(client.byol_virtualization),
         vcpu_per_core_ratio=client.vcpu_per_pcore_ratio or 1.97,
-        pcores_with_virtualization=int(client.pcores_with_virtualization),
-        pcores_with_windows_server=int(client.pcores_with_windows_server),
-        pcores_with_windows_esu=int(client.pcores_with_windows_server_esu),
-        pcores_with_sql_server=int(client.pcores_with_sql_server),
-        pcores_with_sql_esu=int(client.pcores_with_sql_server_esu),
+        # Pass fractional pCore counts verbatim — BA workbook hand-types
+        # fractional values (e.g., =12405/1.48 = 8381.756...) and the per-core
+        # licensing rate applies to the fractional value, not a rounded int.
+        pcores_with_virtualization=float(client.pcores_with_virtualization),
+        pcores_with_windows_server=float(client.pcores_with_windows_server),
+        pcores_with_windows_esu=float(client.pcores_with_windows_server_esu),
+        pcores_with_sql_server=float(client.pcores_with_sql_server),
+        pcores_with_sql_esu=float(client.pcores_with_sql_server_esu),
     )
 
     # ---- Consumption plan ----
